resource_group: fix cherry-pick of #11023
