+++ client/resource_group/controller/controller_test.go
@@ -21,6 +21,7 @@ package controller
 import (
 	"context"
 	"fmt"
+	"math"
 	"testing"
 	"time"
 
@@ -69,6 +70,31 @@ func TestGroupControlBurstable(t *testing.T) {
 	}
 	gc.updateAvgRequestResourcePerSec()
 	re.True(gc.burstable.Load())
+}
+
+func TestCalcAvgSkipsNonPositiveDeltaDuration(t *testing.T) {
+	re := require.New(t)
+	gc := createTestGroupCostController(re)
+	counter := gc.run.requestUnitTokens[rmpb.RequestUnitType_RU]
+	// Simulate a controller that is published between the `updateRunState` and
+	// `updateAvgRequestResourcePerSec` passes of the same tick: `gc.run.now`
+	// still equals the `avgLastTime` set by `initRunState`, so the elapsed
+	// duration is exactly zero and the sample must be skipped instead of
+	// computing 0/0 = NaN.
+	re.Equal(gc.run.now, counter.avgLastTime)
+	gc.updateAvgRequestResourcePerSec()
+	re.False(math.IsNaN(counter.avgRUPerSec))
+	re.Zero(counter.avgRUPerSec)
+	// The skipped sample must not advance the accounting state.
+	re.Equal(gc.run.now, counter.avgLastTime)
+	re.Zero(counter.avgRUPerSecLastRU)
+
+	// Once the run state advances, sampling resumes as usual.
+	gc.run.consumption.RRU = 100
+	gc.run.now = gc.run.now.Add(time.Second)
+	gc.updateAvgRequestResourcePerSec()
+	re.False(math.IsNaN(counter.avgRUPerSec))
+	re.Positive(counter.avgRUPerSec)
 }
 
 func TestRequestAndResponseConsumption(t *testing.T) {
